requirements.txt
Flask==2.3.2
python-dotenv==1.0.0
Werkzeug==2.3.6
gunicorn==20.1.0
